pyproject.toml
[tool.poetry]
name = "gptdemo"
version = "0.1.0"
description = ""
authors = ["nahrun1682 <kawamura100995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.26.0"
langchain = "^0.0.285"
openai = "^0.28.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
